Add files via upload
The R code creates the brazilEcon dataframe. First, csv files are loaded and joined into the byMonth, byQuarter, and byYear dataframes. The byQuarter and byYear dataframes are modfied to create 36 rows by duplicating existing rows (3 times for byQuarter and 12 times for byYear), and updating the months to values for each month. The byMonth, byQuarter, and byYear dataframes are joined, and saved as a csv file.